requirements.txt
Flask==3.1.1
line-bot-sdk==3.17.1
openai==1.98.0
requests==2.31.0